separate template tests
